refactor: made new files, deleted unused
