+++ src/components/hooks/useBooks.js
@@ -1,10 +1,9 @@
-import React, { useEffect, useState } from "react"
+import React, { useState } from "react"
+import { userData } from "../../helpers"
 
 export const useBooks = () => {
   const [books, setBooks] = useState([])
-  const tokenString = localStorage.getItem('user');
-  const token = JSON.parse(tokenString);
-  // const mytransactionApiUrl = 'api/users/me?populate[transactions][populate][book][populate][book][populate]=*'
+  const token = userData()
   const openTransactionsQuerry = 'api/users/me?populate[transactions][populate][book][populate][book][populate]=*&populate[transactions][filters][open]=true'
 
   
@@ -19,10 +18,6 @@ export const useBooks = () => {
         setBooks(book.transactions);
        });
   }
-
-  // useEffect(() => {
-  //   fetchBooks()
-  // }, [])
 
   const updateBooks = () => {
     fetchBooks();
